+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Сборка переведена на yamoney-gradle-project-plugin=5.+
+* Обновлён yamoney-java-module-plugin
 ### NEXT_VERSION_DESCRIPTION_END
 ## [3.4.4]() (08-05-2019)